Move _set_fallbacks to standalone function
